# web/openapi.json
{
  "components": {
    "schemas": {
...
      "AlertPage": {
        "description": "Cursor-paginated, per the \u00a78 convention.\n\n`alert_events` is the only table here that grows without bound \u2014 budgets\nare capped at one per period per project and recommendations are replaced\nnightly, so both are bounded by construction and return plain lists. This\none is a history, and a year-old incident is still a row. See ADR 0009.",
        "properties": {
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/AlertResponse"
            },
            "title": "Alerts",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "alerts",
          "next_cursor"
        ],
        "title": "AlertPage",
        "type": "object"
      },
      "AlertResponse": {
        "properties": {
          "alert_type": {
            "title": "Alert Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "detail": {
            "additionalProperties": true,
            "title": "Detail",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "notified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notified At"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "resolution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "alert_type",
          "severity",
          "title",
          "detail",
          "status",
          "notified_at",
          "resolved_at",
          "resolution",
          "created_at"
        ],
        "title": "AlertResponse",
        "type": "object"
      },
...
      "BudgetResponse": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "fraction_used": {
            "title": "Fraction Used",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "limit_usd": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Limit Usd",
            "type": "string"
          },
          "period": {
            "title": "Period",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "spent_usd": {
            "title": "Spent Usd",
            "type": "number"
          }
        },
        "required": [
          "id",
          "project_id",
          "period",
          "limit_usd",
          "action",
          "is_active",
          "spent_usd",
          "fraction_used",
          "created_at"
        ],
        "title": "BudgetResponse",
        "type": "object"
      },
...
      "CreateBudgetRequest": {
        "properties": {
          "action": {
            "default": "alert_only",
            "enum": [
              "alert_only",
              "soft_throttle",
              "hard_stop"
            ],
            "title": "Action",
            "type": "string"
          },
          "limit_usd": {
            "anyOf": [
              {
                "minimum": 1e-06,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Limit Usd"
          },
          "period": {
            "enum": [
              "daily",
              "weekly",
              "monthly"
            ],
            "title": "Period",
            "type": "string"
          },
          "project_id": {
            "minLength": 1,
            "title": "Project Id",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "period",
          "limit_usd"
        ],
        "title": "CreateBudgetRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KillSwitchResponse": {
        "properties": {
          "alert_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alert Id"
          },
          "keys_revoked": {
            "title": "Keys Revoked",
            "type": "integer"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "took_ms": {
            "title": "Took Ms",
            "type": "number"
          }
        },
        "required": [
          "project_id",
          "keys_revoked",
          "took_ms",
          "alert_id"
        ],
        "title": "KillSwitchResponse",
        "type": "object"
      },
...
      "ResolveAlertRequest": {
        "properties": {
          "resolution": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          },
          "status": {
            "default": "resolved",
            "enum": [
              "acknowledged",
              "resolved"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "title": "ResolveAlertRequest",
        "type": "object"
      },
...
      "UpdateBudgetRequest": {
        "properties": {
          "action": {
            "anyOf": [
              {
                "enum": [
                  "alert_only",
                  "soft_throttle",
                  "hard_stop"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "limit_usd": {
            "anyOf": [
              {
                "minimum": 1e-06,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Usd"
          }
        },
        "title": "UpdateBudgetRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Control plane: accounts, projects, keys, usage, and advice.",
    "title": "APICost Dashboard API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/alert-events": {
      "get": {
        "description": "Alert history, newest first \u2014 UC-34.",
        "operationId": "list_alerts_alert_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Alerts",
        "tags": [
          "budgets"
        ]
      }
    },
    "/alert-events/{alert_id}/resolve": {
      "post": {
        "description": "Record what the user did about an alert (UC-34).\n\nThe resolution note is the part worth keeping. \"Was this real, and what did\nwe do\" is the question a user asks of an alert history six weeks later, and\na status flag alone cannot answer it.",
        "operationId": "resolve_alert_alert_events__alert_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveAlertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resolve Alert",
        "tags": [
          "budgets"
        ]
      }
    },
...
    "/budgets": {
      "get": {
        "operationId": "list_budgets_budgets_get",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BudgetResponse"
                  },
                  "title": "Response List Budgets Budgets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Budgets",
        "tags": [
          "budgets"
        ]
      },
      "post": {
        "operationId": "create_budget_budgets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBudgetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Budget",
        "tags": [
          "budgets"
        ]
      }
    },
    "/budgets/{budget_id}": {
      "delete": {
        "operationId": "delete_budget_budgets__budget_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "budget_id",
            "required": true,
            "schema": {
              "title": "Budget Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Budget",
        "tags": [
          "budgets"
        ]
      },
      "patch": {
        "operationId": "update_budget_budgets__budget_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "budget_id",
            "required": true,
            "schema": {
              "title": "Budget Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBudgetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Budget",
        "tags": [
          "budgets"
        ]
      }
    },
...
    "/projects/{project_id}/kill": {
      "post": {
        "description": "Revoke every proxy key for a project, immediately (UC-33).\n\nMust take effect in under a second (BUILD_SPEC \u00a74 P6), which rules out\nwaiting for the 60 s auth cache to expire \u2014 so the revocation and the cache\npurge happen together, and the purge is what actually makes it instant.\n\nProvider keys are deliberately untouched. The user is containing a leak of\n*our* credential; destroying their OpenAI key as well would turn one\nincident into two, and it is not ours to destroy.",
        "operationId": "kill_project_projects__project_id__kill_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KillSwitchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Kill Project",
        "tags": [
          "budgets"
        ]
      }
    },
...
  }
}